requirements.txt
airtable-python-wrapper==0.15.1
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.3
wrapt==1.12.1
